refactor(consistent filenames): use config.js instead of conf.js
